Refactor. Test remove multiple items
